SPARKC-4: Refactoring: Remove ClassTag from ColumnMapper.
It is not used by most ColumnMapper implementations, so there is no need to force it in the base class. Make Java API obtain a ClassTag from Class on the fly, when needed, instead of taking it from a ColumnMapper.

It was used only by JavaBeanColumnMapper. There is no point in forcing all the implementations to require it. Most implementations use TypeTag, which provides stronger type information than ClassTag.