SPARKC-4: Introduce GettableDataToMappedTypeConverter.
GettableDataToMappedTypeConverter can create an object of class T
from a GettableData instance, e.g. a CassandraRow or a UDTValue,
using the column mapping provided by an appropriate ColumnMapper[T].
It is ok to nest UDTs inside of UDTs as well as inside of collections,
because object fields and collection items are converted recursively.

This commit only introduces the new type converter with unit tests,
but the class is not yet plugged into the rest of the system, so
row readers still don't know how to convert UDTs. This will be addressed
by the next commit.